requirements.txt
django
requests
django-rest-framework
bs4
python-dateutil
feedparser
django-environ
pytest
pytest-django